teams_bot/requirements.txt
# Runtime dependencies
azure-functions>=1.12.0
botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0
botframework-connector>=4.14.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
opencensus-ext-azure>=1.1.0
opencensus-ext-logging>=0.1.0
opencensus-ext-requests>=0.8.0
msrest>=0.7.0

# Development dependencies
black>=22.3.0
pylint>=2.15.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
python-dotenv>=1.0.0
rdflib>=6.2.0
pyshacl>=0.20.0
